Use normal array instead of std::vector, use float instead of double

+++ src/include/match_scorers.h
@@ -12,18 +12,18 @@ class MatchScorer {
           \param el2 an element from the second sequence
           \return alignment score for given elements
         */
-        virtual double getScore(char el1, char el2) const = 0;
+        virtual float getScore(char el1, char el2) const = 0;
         virtual ~MatchScorer() {}
 };
 
 /*! A match score is the score of identical elements, otherwise mismatch score */
 class ConstMatchScorer : public MatchScorer {
     private:
-        double matchScore;
-        double mismatchScore;
+        float matchScore;
+        float mismatchScore;
     public:
-        ConstMatchScorer(double, double);
-        double getScore(char el1, char el2) const;
+        ConstMatchScorer(float, float);
+        float getScore(char el1, char el2) const;
         ConstMatchScorer() {}
 };
 
@@ -31,20 +31,20 @@ class ConstMatchScorer : public MatchScorer {
 class DictMatchScorer : public MatchScorer {
     private:
         std::map<char, int> elementsIndices;
-        double** scoreMatrix;
+        float** scoreMatrix;
     public:
-        DictMatchScorer(std::map<char, int>, double**);
-        double getScore(char el1, char el2) const;
+        DictMatchScorer(std::map<char, int>, float**);
+        float getScore(char el1, char el2) const;
         DictMatchScorer() {}
 };
 
 /*! A match score is a result of a callback function */
 class CallbackMatchScorer : public MatchScorer {
     private:
-        double(*scoreCallback)(char, char);
+        float(*scoreCallback)(char, char);
     public:
-        CallbackMatchScorer(double(*)(char, char));
-        double getScore(char el1, char el2) const;
+        CallbackMatchScorer(float(*)(char, char));
+        float getScore(char el1, char el2) const;
         CallbackMatchScorer() {}
 };
